feat: Allow changing units of measure for reporting
Added "default_units" to the config file.  This globally
overrides/forces the setting which can be: metric(m) US(u) or metric except wind(M).

These values may also be specified at the end of a `!weather location`
call from a chat prefixed by 'u:" to indicate you are specifying the unit to use.
So you can do something like:

!weather Chicago, Il u:m

To get the weather in Chicago reported in Celcius.

closes #3